For the past two weeks I've been procrastinating from fixing a double-free bug in the JS accessor for Ship.escorts (or, more precisely, in the code constructing a JS array). The bug was in r1129, so not checking stuff in doesn't really make much sense. Fixes in that time: * Clean, integrated model for JS ship scripts and world scripts. * Player ships support JS ship scripts (but not ones synthesized from legacy actions). * Subentity shader uniform bindings bind to owning ship as intended. * Fixed like_ship recursion limiter in -[Universe getDictionaryForShip:]. * Fixed a bug in material configuration parsing. * Fixed a material bug where a shader material with textures, followed by a shader material with no textures, followed by a basic material with no textures would leave a texture bound. * Cleaned up NSArray -> JS array conversion, without fixing aforementioned bug. * Made JS Ship property "target" read/write, and made property setting work. * Added JS Quaternion.random() static method.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@1136 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Cocoa/MyOpenGLView.m
@@ -78,7 +78,7 @@ static NSString * kOOLogKeyDown				= @"input.keyMapping.keyPress.keyDown";
         NSOpenGLPFAAccelerated,
         0
     };
-    long rendererID;
+    GLint rendererID;
 
     // Create our non-FullScreen pixel format.
     NSOpenGLPixelFormat* pixelFormat = [[NSOpenGLPixelFormat alloc] initWithAttributes:attrs];
